postgresql: move PgMembership to module_utils
